requirements.txt
pymysql
python-dotenv
beautifulsoup4
selenium